+++ src/main/java/insane96mcp/progressivebosses/mixin/EndCrystalMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.progressivebosses.mixin;
 
-import insane96mcp.progressivebosses.module.dragon.data.DragonCrystal;
+import insane96mcp.progressivebosses.module.dragon.data.SpikesComponent;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.boss.enderdragon.EndCrystal;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,7 +13,7 @@ public class EndCrystalMixin {
 
 	@Inject(at = @At("HEAD"), method = "hurt(Lnet/minecraft/world/damagesource/DamageSource;F)Z", cancellable = true)
 	private void hurt(DamageSource source, float amount, CallbackInfoReturnable<Boolean> callback) {
-		if (DragonCrystal.onCrystalDamagedByExplosion(source))
+		if (SpikesComponent.onCrystalDamagedByExplosion(source))
 			callback.setReturnValue(false);
 	}
 }